Started work on Sprite class

+++ enginev2/graphics/commons/primitives/Dimensions.cpp
@@ -1,10 +1,6 @@
 #include "Dimensions.h"
 
-Dimensions::Dimensions() :
-		w(0), h(0) {
-}
-Dimensions::Dimensions(int32_t width, int32_t height) :
-		w(width), h(height) {
-}
+Dimensions::Dimensions(): w(0), h(0) {}
+Dimensions::Dimensions(uint16_t width, uint16_t height): w(width), h(height) {}
 
 const Dimensions Dimensions::UNDEFINED(0, 0);
